feat: add 'coupon-infra' module and move on api to infra module
