+++ coupon/coupon-api/src/main/java/dev/be/coupon/api/coupon/application/CouponService.java
@@ -9,17 +9,17 @@ import dev.be.coupon.api.coupon.application.dto.CouponUsageResult;
 import dev.be.coupon.api.coupon.application.exception.CouponIssueException;
 import dev.be.coupon.api.coupon.application.exception.CouponNotFoundException;
 import dev.be.coupon.api.coupon.application.exception.IssuedCouponNotFoundException;
-import dev.be.coupon.api.coupon.infrastructure.kafka.producer.CouponIssueProducer;
-import dev.be.coupon.api.coupon.infrastructure.redis.AppliedUserRepository;
-import dev.be.coupon.api.coupon.infrastructure.redis.CouponCacheRepository;
-import dev.be.coupon.api.coupon.infrastructure.redis.CouponCountRedisRepository;
-import dev.be.coupon.api.coupon.infrastructure.redis.aop.DistributedLock;
 import dev.be.coupon.domain.coupon.Coupon;
 import dev.be.coupon.domain.coupon.CouponRepository;
 import dev.be.coupon.domain.coupon.IssuedCoupon;
 import dev.be.coupon.domain.coupon.IssuedCouponRepository;
 import dev.be.coupon.domain.coupon.UserRoleChecker;
 import dev.be.coupon.domain.coupon.exception.UnauthorizedAccessException;
+import dev.be.coupon.infra.kafka.producer.CouponIssueProducer;
+import dev.be.coupon.infra.redis.AppliedUserRepository;
+import dev.be.coupon.infra.redis.CouponCacheRepository;
+import dev.be.coupon.infra.redis.CouponCountRedisRepository;
+import dev.be.coupon.infra.redis.aop.DistributedLock;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.kafka.KafkaException;
