+++ coupon/coupon-api/src/test/java/dev/be/coupon/api/coupon/application/CouponServiceTest.java
@@ -9,16 +9,16 @@ import dev.be.coupon.api.coupon.application.dto.CouponUsageResult;
 import dev.be.coupon.api.coupon.application.exception.CouponIssueException;
 import dev.be.coupon.api.coupon.application.exception.InvalidIssuedCouponException;
 import dev.be.coupon.api.coupon.application.exception.IssuedCouponNotFoundException;
-import dev.be.coupon.api.coupon.infrastructure.kafka.producer.CouponIssueProducer;
-import dev.be.coupon.api.coupon.infrastructure.redis.AppliedUserRepository;
-import dev.be.coupon.api.coupon.infrastructure.redis.CouponCacheRepository;
-import dev.be.coupon.api.coupon.infrastructure.redis.CouponCountRedisRepository;
 import dev.be.coupon.domain.coupon.CouponRepository;
 import dev.be.coupon.domain.coupon.IssuedCoupon;
 import dev.be.coupon.domain.coupon.IssuedCouponRepository;
 import dev.be.coupon.domain.coupon.exception.CouponAlreadyUsedException;
 import dev.be.coupon.domain.coupon.exception.InvalidCouponTypeException;
 import dev.be.coupon.domain.coupon.exception.UnauthorizedAccessException;
+import dev.be.coupon.infra.kafka.producer.CouponIssueProducer;
+import dev.be.coupon.infra.redis.AppliedUserRepository;
+import dev.be.coupon.infra.redis.CouponCacheRepository;
+import dev.be.coupon.infra.redis.CouponCountRedisRepository;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.awaitility.Awaitility.await;
