+++ coupon/coupon-infra/src/main/java/dev/be/coupon/infra/kafka/dto/CouponIssueMessage.java
@@ -1,0 +1,6 @@
+package dev.be.coupon.infra.kafka.dto;
+
+import java.util.UUID;
+
+public record CouponIssueMessage(UUID userId, UUID couponId) {
+}
