+++ coupon/coupon-kafka-consumer/src/test/resources/application-test.yml
@@ -10,7 +10,11 @@ spring:
       hibernate:
         format_sql: true
         show_sql: true
-
+  # sprig boot 3.x
+  data:
+    redis:
+      host: localhost
+      port: 6379
   kafka:
     producer:
       bootstrap-servers: localhost:9092
